various fixes

+++ src/el1/io_text_parser.cpp
@@ -3,7 +3,7 @@
 
 namespace el1::io::text::parser
 {
-	TParser<ast::TLiteralNode> operator""_P(const char* const str, unsigned long len)
+	TParser<ast::TLiteralNode> operator""_P(const char* const str, size_t len)
 	{
 		return TParser(ast::TLiteralNode(TString(str, len)));
 	}
